fix: group layers, remove deep omit to have precise types, and move types
